Code changes: * Major AI reorganization - Added Task and Scheduler classes based on Game::AI++ by Alex J. Champandard - Added Behavior base class that keeps an owner ID and points to its parent Controller - Added Brain class that contains a scheduler - Split Aimer behaviors out into separate Behaviors - Added Tag component representing data attached to entities by or for the AI * Added support for boundary violation listeners - Renamed Collidable::Listener as Collidable::ContactListener - Added Collidable::BoundaryListener * Added GetId to Controller * Added PostActivate and PreDeactivate component initialization events to break order dependencies * Replaced glVertex4 and glTexCoord4 with glVertex3 and glTexCoord2 to simplify drawlists * Changed Entity::GetTransform to return a const reference * Added LinkTemplate flags to control linked-entity deletion and team propagation * PlayerController activates aiming reticule when using AIMCURSOR or MOVECURSOR * HACK: PlayerHUD switches game to shell after Game Over display has been active for long enough * Escape menu pauses audio * Sound system music enhancements - Split SOUND_VOLUME into SOUND_VOLUME_EFFECT and SOUND_VOLUME_MUSIC - Added a proper component initializer for music * Added Spawn class encapsulating spawn behavior * Removed obsolete element value check from Spawner::Configure * Added title from Nova Core * Added WaveSequence class from Nova Core representing sequenced attack waves

+++ Source/Link.cpp
@@ -24,7 +24,6 @@ namespace Database
 	Typed<Typed<LinkTemplate> > linktemplate(0x801f01af /* "linktemplate" */);
 	Typed<Typed<Link *> > link(0x0ddb0669 /* "link" */);
 	Typed<unsigned int> backlink(0xe3736e9a /* "backlink" */);
-	Typed<bool> below(0x9253c0f2 /* "below" */);
 
 	namespace Loader
 	{
@@ -120,7 +119,8 @@ LinkTemplate::LinkTemplate(void)
 , mType(0)
 , mUpdateAngle(true)
 , mUpdatePosition(true)
-, mBelow(false)
+, mUpdateTeam(true)
+, mDeleteSecondary(true)
 {
 }
 
@@ -151,10 +151,15 @@ bool LinkTemplate::Configure(const TiXmlElement *element, unsigned int aId, unsi
 	element->QueryIntAttribute("updateposition", &updateposition);
 	mUpdatePosition = updateposition != 0;
 
-	// relative depth
-	int below = mBelow;
-	element->QueryIntAttribute("below", &below);
-	mBelow = below != 0;
+	// update linked team?
+	int updateteam = mUpdateTeam;
+	element->QueryIntAttribute("updateteam", &updateteam);
+	mUpdateTeam = updateteam != 0;
+
+	// delete linked secondary?
+	int deletesecondary = mDeleteSecondary;
+	element->QueryIntAttribute("deletesecondary", &deletesecondary);
+	mDeleteSecondary = deletesecondary != 0;
 
 	// custom template identifier (if any)
 	unsigned int custom = 0U;
@@ -230,19 +235,19 @@ Link::Link(const LinkTemplate &aTemplate, unsigned int aId)
 			transform.Angle(), transform.p, entity->GetVelocity(), entity->GetOmega(), false);
 	}
 
-	// if the owner has a team...
-	unsigned int team = Database::team.Get(mId);
-	if (team)
+	if (aTemplate.mUpdateTeam)
 	{
-		// propagate team to spawned item
-		Database::team.Put(mSecondary, team);
+		// if the owner has a team...
+		unsigned int team = Database::team.Get(mId);
+		if (team)
+		{
+			// propagate team to spawned item
+			Database::team.Put(mSecondary, team);
+		}
 	}
 
 	// create a backlink
 	Database::backlink.Put(mSecondary, mId);
-
-	// propagate below
-	Database::below.Put(mSecondary, aTemplate.mBelow);
 
 	if (!Database::entity.Get(aTemplate.mSecondary))
 	{
@@ -255,7 +260,14 @@ Link::~Link(void)
 {
 	if (mSecondary)
 	{
-		Database::Delete(mSecondary);
+		// get template data
+		const LinkTemplate &link = Database::linktemplate.Get(mId).Get(mSub);
+		
+		// if deleting the secondary
+		if (link.mDeleteSecondary)
+		{
+			Database::Delete(mSecondary);
+		}
 	}
 }
